Create Distance
Write a program to find the distance between 2 points using functions.

Input Format:

Input consists of 4 integers. The first and second integer corresponds to the x-coordinate and y-coordinate of the first point. The third and fourth integer corresponds to the x-coordinate and y-coordinate of the second point.

Output Format:

Output consists of a single floating point number (correct to 2 decimal places.) Refer sample output for formatting details.

Sample Input:

3

6

4

3

Sample Output:

Distance between the 2 points is 3.16