Create ASCII Value
Ravi and Varma went to a textile shop for festival purchase. Ravi took few T-shirts and went for billing. Varma wants discount on the purchased items. The biller shows computer keyboard and asks to press an alpha numeric key and that key’s ASCII value will be the discounted amount on the purchase. Help Varma to know his discounted amount.

Sample Input:

A

Sample Output:

65